// src/gen-ts/src/models/LoginRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni.
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface LoginRequest
 */
export interface LoginRequest {
  /**
   *
   * @type {string}
   * @memberof LoginRequest
   */
  email?: string
  /**
   *
   * @type {string}
   * @memberof LoginRequest
   */
  password?: string
}

/**
 * Check if a given object implements the LoginRequest interface.
 */
export function instanceOfLoginRequest(value: object): value is LoginRequest {
  return true
}

export function LoginRequestFromJSON(json: any): LoginRequest {
  return LoginRequestFromJSONTyped(json, false)
}

export function LoginRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): LoginRequest {
  if (json == null) {
    return json
  }
  return {
    email: json['email'] == null ? undefined : json['email'],
    password: json['password'] == null ? undefined : json['password'],
  }
}

export function LoginRequestToJSON(json: any): LoginRequest {
  return LoginRequestToJSONTyped(json, false)
}

export function LoginRequestToJSONTyped(
  value?: LoginRequest | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    email: value['email'],
    password: value['password'],
  }
}